requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Gmail API
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2

# AI/ML
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.5
sentence-transformers==2.3.1
pinecone==8.1.0
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.26.0
redis==5.0.1
celery==5.3.6

# Frontend
streamlit>=1.31.0
plotly>=5.18.0

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0
